pyproject.toml
[tool.poetry]
name = "oura_analysis"
version = "0.1.0"
description = "Tools to analyse data from Oura."
authors = ["Sten Remmelg <sten.remmelg@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.1.4"
plotly = "^5.18.0"
ipykernel = "^6.28.0"
nbformat = "^5.9.2"
duckdb = "^1.3.2"
requests = "^2.32.5"
python-dotenv = "^1.1.1"
loguru = "^0.7.3"
streamlit = "^1.48.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.12.1"
isort = "^5.13.2"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
